Remove transformation module and incorporate its functionality into transform function in uflnamespace
